Move format file size to utils.
